+++ tests/assets/esbuild/main.ts
@@ -1,6 +1,7 @@
 /// <reference lib="dom" />
 import toUppercase, { toLowercase } from "./modules/to_uppercase.ts";
-import data from "./data.json";
+import { Foo } from "./main.d.ts";
+import data from "./data.json" with { type: "json" };
 
 // https://github.com/lumeland/lume/issues/442
 import "https://esm.sh/v127/prop-types@15.8.1/denonext/prop-types.development.mjs";
@@ -13,3 +14,6 @@ document.querySelectorAll("h1")?.forEach((h1) => {
       h1.innerHTML = lower;
     });
 });
+
+const foo: Foo = { bar: "baz" };
+console.log(foo);