+++ tests/config.test.ts
@@ -44,7 +44,7 @@ Deno.test("ignored files configuration", () => {
   const { ignored, filters } = site.source;
 
   equals(ignored.size, 7);
-  equals(filters.length, 0);
+  equals(filters.length, 1);
   equals(ignored.has("/node_modules"), true);
   equals(ignored.has("/_site"), true);
   equals(ignored.has("/import_map.json"), true);
@@ -52,20 +52,20 @@ Deno.test("ignored files configuration", () => {
 
   site.ignore("README.md");
   equals(ignored.size, 8);
-  equals(filters.length, 0);
+  equals(filters.length, 1);
   equals(ignored.has("/README.md"), true);
 
   site.ignore("file2", "file3", "README.md");
   equals(ignored.size, 10);
-  equals(filters.length, 0);
+  equals(filters.length, 1);
   equals(ignored.has("/file2"), true);
   equals(ignored.has("/file3"), true);
 
   const filter = (path: string) => path.includes("file");
   site.ignore(filter);
   equals(ignored.size, 10);
-  equals(filters.length, 1);
-  equals(filters[0], filter);
+  equals(filters.length, 2);
+  equals(filters[1], filter);
 });
 
 Deno.test("event listener configuration", () => {